refactor(core): Split some core modules into smaller modules
