+++ alumet/src/ffi/metrics.rs
@@ -6,7 +6,7 @@ use crate::{
     measurement::{
         AttributeValue, MeasurementAccumulator, MeasurementBuffer, MeasurementPoint, WrappedMeasurementValue,
     },
-    metrics::RawMetricId,
+    metrics::{def::RawMetricId, registry::MetricRegistry},
     resources::{Resource, ResourceConsumer},
 };
 
@@ -20,7 +20,7 @@ use super::{
 // ====== Metrics ffi ======
 #[no_mangle]
 pub extern "C" fn metric_name<'a>(metric: RawMetricId, ctx: &'a FfiOutputContext) -> AStr<'a> {
-    let metrics: &crate::metrics::MetricRegistry = unsafe { &*ctx.inner }.metrics;
+    let metrics: &MetricRegistry = unsafe { &*ctx.inner }.metrics;
     let name: &str = &metrics.by_id(&metric).unwrap().name;
     AStr::from(name)
 }
