+++ alumet/src/metrics/mod.rs
@@ -1,0 +1,8 @@
+//! Definition and management of metrics.
+
+pub mod def;
+pub mod error;
+pub mod online;
+pub mod registry;
+
+pub use def::{Metric, RawMetricId, TypedMetricId};
