+++ alumet/src/pipeline/mod.rs
@@ -13,9 +13,9 @@
 pub mod builder;
 pub mod control;
 pub mod elements;
-pub mod registry;
+pub mod naming;
 pub mod trigger;
-mod util;
+pub(crate) mod util;
 
 pub use elements::output::Output;
 pub use elements::source::Source;
