+++ alumet/src/pipeline/naming/key.rs
@@ -1,0 +1,7 @@
+//! Keys that allow to control individual elements.
+
+use super::ElementName;
+
+pub struct SourceKey(ElementName);
+pub struct TransformKey(ElementName);
+pub struct OutputKey(ElementName);
